Move element related to type to type_system
